test: moving tests into their own files
