+++ test/querySelector.spec.js
@@ -7,7 +7,7 @@
 /* global it describe */
 
 import { querySelectorAll, querySelector } from '../src/index.js'
-import { assertSelectorEqual, withDom, simplifyElement, simplifyElements } from './utils.js';
+import { assertSelectorEqual, withDom, simplifyElement, simplifyElements } from './utils.js'
 import assert from 'assert'
 import simpleLight1 from './fixtures/simple1/light.html'
 import simpleShadow1 from './fixtures/simple1/shadow.html'
@@ -49,8 +49,6 @@ import nestedSlotsLight7 from './fixtures/nestedSlots7/light.html'
 import nestedSlotsShadow7 from './fixtures/nestedSlots7/shadow.html'
 import unusualSelectorsLight1 from './fixtures/unusualSelectors1/light.html'
 import unusualSelectorsShadow1 from './fixtures/unusualSelectors1/shadow.html'
-import classNamesLight1 from './fixtures/classNames1/light.html'
-import classNamesShadow1 from './fixtures/classNames1/shadow.html'
 
 function testSelectors (lightDom, shadowDom, tests) {
   tests.forEach(({ selector, expected }) => {
@@ -886,110 +884,5 @@ describe('basic test suite', function () {
       { selector: '[data-hashdot="hash\\#\\.dot"]', expected: hashDot },
       { selector: "[data-hashdot='hash\\#\\.dot']", expected: hashDot }
     ])
-  })
-
-  describe('getElementsByClassName', () => {
-    const expected = [
-      {
-        tagName: 'DIV',
-        classList: ['container', 'main']
-      },
-      {
-        tagName: 'SPAN',
-        classList: ['container', 'main', 'outerText']
-      },
-      {
-        tagName: 'SPAN',
-        classList: ['container', 'main', 'innerText']
-      }
-    ]
-
-    it('light DOM - getElementsByClassName', () => {
-      const classNames = 'container main'
-      withDom(classNamesLight1, context => {
-        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
-      })
-    })
-
-    it('shadow DOM - getElementsByClassName', () => {
-      const classNames = 'container main'
-      withDom(classNamesShadow1, context => {
-        assertSelectorEqual('.container.main', getElementsByClassName(classNames, context), expected, true)
-      })
-    })
-
-    it('light DOM - getElementsByClassName with multiple spaces', () => {
-      withDom(classNamesLight1, context => {
-        const classNames = 'container      main'
-        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
-      })
-    })
-
-    it('shadow DOM - getElementsByClassName with multiple spaces', () => {
-      withDom(classNamesShadow1, context => {
-        const classNames = 'container      main'
-        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
-      })
-    })
-
-    it('light DOM - getElementsByClassName with line breaks', () => {
-      withDom(classNamesLight1, context => {
-        const classNames = `main
-        
-        
-        
-        container`
-        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
-      })
-    })
-
-    it('shadow DOM - getElementsByClassName with line breaks', () => {
-      withDom(classNamesShadow1, context => {
-        const classNames = `main
-        
-        
-        
-        container`
-        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
-      })
-    })
-
-    /* eslint-disable no-tabs */
-
-    it('light DOM - getElementsByClassName with tabs', () => {
-      withDom(classNamesLight1, context => {
-        const classNames = 'main	container'
-        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
-      })
-    })
-
-    it('shadow DOM - getElementsByClassName with tabs', () => {
-      withDom(classNamesShadow1, context => {
-        const classNames = 'main	container'
-        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
-      })
-    })
-
-    it('light DOM - getElementsByClassName with spaces, line breaks, and tabs', () => {
-      withDom(classNamesLight1, context => {
-        const classNames = `	main	container
-        
-            	
-        `
-        assertSelectorEqual(classNames, context.getElementsByClassName(classNames), expected, true)
-      })
-    })
-
-    it('shadow DOM - getElementsByClassName with spaces, line breaks, and tabs', () => {
-      withDom(classNamesShadow1, context => {
-        const classNames = `	main	container
-        
-            	
-        `
-        assertSelectorEqual(classNames, getElementsByClassName(classNames, context), expected, true)
-      })
-    })
-
-    /* eslint-enable no-tabs */
   })
 })
